ui: rearrange route modules
auth is kinda its own thing so lets split that out
